Unify color components
